backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
python-dotenv==1.0.1
requests==2.31.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.120.0
google-auth-httplib2==0.2.0
Authlib==1.3.0
gunicorn==21.2.0
psycopg2==2.9.10
Flask-JWT-Extended==4.5.2
